+++ CHANGELOG.md
@@ -10,7 +10,17 @@ Compatibility:
 Features:
   - Upgrade to Decidim v0.30
 
-v0.12.2
+v0.12.4
+-------
+Compatibility:
+  - Decidim 0.29.x
+
+Features:
+  - Added application context for config constraints (whether the user is logged or not)
+  - Allow to granular permissions to forced verifications.
+    **NOTE** Config var `force_authorization_after_login` has changed to `force_authorizations`
+
+v0.12.3
 -------
 
 Compatibility: